tried simpliflying to get channels to work
