--- core/config.py
@@ -1,5 +1,0 @@
-import os
-
-
-class Config:
-    SECRET_KEY = os.getenv('SECRET_KEY')
