--- core/templates/index.html
@@ -1,31 +1,0 @@
-{% extends "layout.html" %}
-
-{% block content %}
-<form method="POST" action="" class="content-section form-inline hidden" id="username-form">
-    <div class="form-group mx-sm-3">
-        <input class="form-control mr-3" id="username" autocomplete="off" autofocus placeholder="Enter Username" type="text">
-        <input class="btn btn-outline-secondary" id="submit" type="submit" value="Submit">
-    </div>
-</form>
-<form action="" class="content-section form-inline hidden" id="channelName-form">
-    <div class="form-group mx-sm-3">
-        <input class="form-control mr-3" id="channelName" autocomplete="off" autofocus placeholder="Enter Channel Name" type="text">
-        <input class="btn btn-outline-secondary" id="channelSubmit" type="submit" value="Submit">
-    </div>
-</form>
-<br>
-<div id="result"></div>
-<br>
-<div class="mt-5" style="text-align:left;">
-    {% if channels["total"] > 0 %}
-        <p>Number of Channels: <span id="channels">{{ channels["total"] }}</span></p>
-    {% endif %}
-</div>
-<div>
-    <ul id="channels-list">
-        {% for channelName in channelNames %}
-            <li>{{ channelName }}</li>
-        {% endfor %}
-    </ul>
-</div>
-{% endblock content %}
