+++ static/style.css
@@ -1,6 +1,6 @@
 /* ---------------------------------------------------
     TYPE IN THIS COMMAND TO SETUP:
-        sass core/static/style.scss core/static/style.css
+        sass static/style.scss static/style.css
 ----------------------------------------------------- */
 /* ---------------------------------------------------
     SASS VARIABLES
@@ -66,108 +66,16 @@ a:focus {
   font-size: 0.8em;
 }
 
-img {
-  margin-left: 40px;
-}
-
-#headshot {
-  max-width: 75%;
-}
-
 .error-message {
   font-size: 0.8em;
   color: red;
 }
 
 /* ---------------------------------------------------
-    SIDEBAR STYLE
------------------------------------------------------ */
-.wrapper {
-  display: flex;
-  width: 100%;
-}
-
-#sidebar {
-  width: 250px;
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 100vh;
-  z-index: 999;
-  background: #fdaaaa;
-  color: black;
-  transition: all 0.3s;
-}
-
-#sidebar.active {
-  margin-left: -250px;
-}
-
-/* SCSS Nesting */
-#sidebar .sidebar-header {
-  padding-left: 10px;
-  padding-top: 20px;
-  background: #fa6d6d;
-  min-height: 90px;
-}
-#sidebar ul.components {
-  padding: 20px 0;
-  border-bottom: 1px solid #47748b;
-}
-#sidebar ul p {
-  color: black;
-  padding: 10px;
-}
-#sidebar ul li a {
-  padding: 10px;
-  font-size: 1.1em;
-  display: block;
-}
-#sidebar ul li a:hover {
-  color: #fdaaaa;
-  background: #fff;
-}
-#sidebar ul li.active > a,
-#sidebar a[aria-expanded=true] {
-  color: black;
-  background: #fa6d6d;
-}
-
-a[data-toggle=collapse] {
-  position: relative;
-}
-
-.dropdown-toggle::after {
-  display: block;
-  position: absolute;
-  top: 50%;
-  right: 20px;
-  transform: translateY(-50%);
-}
-
-ul ul a {
-  font-size: 0.9em !important;
-  padding-left: 30px !important;
-  background: #fdaaaa;
-}
-
-ul.CTAs {
-  padding: 20px;
-}
-
-ul.CTAs a {
-  text-align: center;
-  font-size: 0.9em !important;
-  display: block;
-  border-radius: 5px;
-  margin-bottom: 5px;
-}
-
-/* ---------------------------------------------------
     CONTENT STYLE
 ----------------------------------------------------- */
 #content {
-  width: calc(100% - 250px);
+  width: 100%;
   padding: 0px;
   min-height: 100vh;
   transition: all 0.3s;
@@ -195,30 +103,6 @@ ul.CTAs a {
   margin-bottom: 20px;
 }
 
-/* ---------------------------------------------------
-    MEDIAQUERIES
------------------------------------------------------ */
-@media (max-width: 768px) {
-  #sidebar {
-    margin-left: -250px;
-  }
-
-  #sidebar.active {
-    margin-left: 0;
-  }
-
-  #content {
-    width: 100%;
-  }
-
-  #content.active {
-    width: calc(100% - 250px);
-  }
-
-  #sidebarCollapse span {
-    display: none;
-  }
-}
 /* ---------------------------------------------------
     MODAL POP-UP
 ----------------------------------------------------- */
